Move CodeInfo into test_record and use it in more places
